now using @string, added box around inquiry multiline text
