+++ LawnCare/frontend/app/src/main/res/layout/fragment_contact.xml
@@ -24,9 +24,8 @@
                 android:layout_height="wrap_content"
                 android:layout_marginHorizontal="20dp"
                 android:layout_marginTop="20dp"
-                android:background="#FFFFFF"
                 android:gravity="center"
-                android:text="Contact Us!" />
+                android:text="@string/contact_title" />
 
             <TextView
                 android:id="@+id/fNameLabel"
@@ -34,7 +33,7 @@
                 android:layout_height="wrap_content"
                 android:layout_marginHorizontal="20dp"
                 android:layout_marginTop="40dp"
-                android:text="First Name"
+                android:text="@string/contact_name_label"
                 app:layout_constraintBottom_toTopOf="@+id/fNameInput" />
 
             <EditText
@@ -45,33 +44,16 @@
                 android:layout_marginHorizontal="20dp"
                 android:layout_marginTop="5dp"
                 android:ems="10"
-                android:hint="First Name" />
+                android:hint="@string/contact_name_hint" />
+
 
             <TextView
-                android:id="@+id/lNameLabel"
+                android:id="@+id/phoneNumberLabel"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginHorizontal="20dp"
                 android:layout_marginTop="15dp"
-                android:text="Last Name" />
-
-            <EditText
-                android:id="@+id/lNameInput"
-                android:layout_width="match_parent"
-                android:layout_height="40dp"
-                android:layout_marginHorizontal="20dp"
-                android:layout_marginTop="5dp"
-                android:ems="10"
-                android:hint="Last Name"
-                android:inputType="textPersonName" />
-
-            <TextView
-                android:id="@+id/phoneNumberLable"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginHorizontal="20dp"
-                android:layout_marginTop="15dp"
-                android:text="Home Phone Number" />
+                android:text="@string/contact_phoneNumber_label" />
 
             <EditText
                 android:id="@+id/phoneNumberInput"
@@ -80,7 +62,7 @@
                 android:layout_marginHorizontal="20dp"
                 android:layout_marginTop="5dp"
                 android:ems="10"
-                android:hint="123-456-7890"
+                android:hint="@string/contact_phoneNumber_hint"
                 android:inputType="phone" />
 
             <TextView
@@ -89,7 +71,7 @@
                 android:layout_height="wrap_content"
                 android:layout_marginHorizontal="20dp"
                 android:layout_marginTop="15dp"
-                android:text="Email" />
+                android:text="@string/contact_email_label" />
 
             <EditText
                 android:id="@+id/emailInput"
@@ -98,7 +80,7 @@
                 android:layout_marginHorizontal="20dp"
                 android:layout_marginTop="5dp"
                 android:ems="10"
-                android:hint="john123@example.com"
+                android:hint="@string/contact_email_hint"
                 android:inputType="textEmailAddress" />
 
             <TextView
@@ -107,7 +89,7 @@
                 android:layout_height="wrap_content"
                 android:layout_marginHorizontal="20dp"
                 android:layout_marginTop="15dp"
-                android:text="Inquiry" />
+                android:text="@string/contact_inquiry_label" />
 
             <EditText
                 android:id="@+id/inquiryInput"
@@ -116,8 +98,9 @@
                 android:layout_marginHorizontal="20dp"
                 android:layout_marginTop="5dp"
                 android:ems="10"
-                android:hint="Ask Us Anything"
+                android:hint="@string/contact_inquiry_hint"
                 android:gravity="start|top"
+                android:background="@drawable/back"
                 android:inputType="textMultiLine" />
             <Button
                 android:id="@+id/submitBtn"
@@ -126,7 +109,7 @@
                 android:layout_marginHorizontal="20dp"
                 android:layout_marginTop="30dp"
                 android:layout_marginBottom="30dp"
-                android:text="Submit" />
+                android:text="@string/submit" />
 
 
         </LinearLayout>
