+++ LawnCare/frontend/app/src/main/res/values/strings.xml
@@ -1,4 +1,5 @@
 <resources>
+    <string name="submit">Submit</string>
     <string name="app_name">LawnCare</string>
     <string name="action_settings">Settings</string>
     <string name="tab_contact">Contact</string>
@@ -9,4 +10,15 @@
     <string name="action_address">Address</string>
     <string name="action_name">Name</string>
     <string name="action_phone">Phone Number</string>
+    <string name="contact_inquiry_hint">Ask Us Anything</string>
+    <string name="contact_inquiry_label">Inquiry</string>
+    <string name="contact_email_hint">john123@example.com</string>
+    <string name="contact_email_label">Email</string>
+    <string name="contact_phoneNumber_hint">123-456-7890</string>
+    <string name="contact_phoneNumber_label">Phone Number</string>
+    <string name="contact_name_hint">Name</string>
+    <string name="contact_name_label">Name</string>
+    <string name="contact_title">Contact Us!</string>
+
+
 </resources>
